♻️ MIN-140 Phase 0: extract sub-agent controller module
Split monolithic orchestrator.ts into controller/{registry,scheduler,controller,watchdog}.
Keep orchestrator.ts as a thin re-export shim for backward compatibility.
Move self-healing/controller watchdog stub to controller/watchdog.ts.
Update context.md to document the new layout and MIN-140 roadmap.
