+++ documentation/context.md
@@ -485,7 +485,7 @@ Drop-in **local** tools without MCP: `~/.minnow/tools/<pluginId>/{tool.json,hand
 
 ### Manual Orchestrate board (parse-only LLM + human-operated Kanban)
 
-Orchestrate mode is **parse-only** for the LLM: read the plan, call **`board_init`** once (per-task `build` / `test`), then stop. Execution is **manual** on the board and in linked **task chats** (no supervisor, no `report_orchestrator_status`, no play/pause Resume).
+Orchestrate mode is **parse-only** for the LLM: read the plan, call **`board_init`** once (per-task `build` / `test`), then stop. Execution is **manual** on the board and in linked **task chats** (no automated supervisor yet; auto-pilot is planned in [MIN-140](plans/Orchestrator%20Plan.md) Phase 3–4).
 
 | Module | Role |
 |--------|------|
@@ -874,7 +874,7 @@ When the active chat has **no messages**, `renderChatFromHistory()` paints the *
 
 Parent tool loop can spawn **isolated sub-agents** (separate messages, model, tool subset). Results return as JSON aggregate tool results; child transcripts are **not** appended to parent `chat.history`.
 
-**Visibility (feature 30):** Each spawn shows a **sub-agent card** in the parent chat (`src/ui/sub-agent-cards.ts`) with live status; clicking opens a **slide-over drawer** with a **structured outcome** (summary, findings, artifacts) and a collapsible read-only transcript (`src/ui/sub-agent-drawer.ts`, `src/styles/sub-agent-drawer.css`). While a run is active, `src/agents/sub-agent-runner.ts` pushes transcript snapshots through `onMessagesChange` (system/user seed, streaming assistant deltas, tool rounds); `src/agents/orchestrator.ts` copies them onto `run.messages` and emits `sub-agent-events`; the open drawer re-renders on each emit. **`list_sub_agents`** / **`get_sub_agent_status`** are **session-scoped** (any parent turn in the same chat; merged with `chat.subAgentRuns` after reload). Terminal runs are copied into `chat.subAgentRuns` (`PersistedSubAgentRun[]` in `src/types.ts`, including `structuredOutcome` / `budgetEvents`) via `src/state/sub-agent-session-sync.ts` so the drawer works after reload. Spawn rows anchor after the parent tool bubble using `data-tool-call-id` on `.tool-call-msg` (`src/tools/loop.ts`).
+**Visibility (feature 30):** Each spawn shows a **sub-agent card** in the parent chat (`src/ui/sub-agent-cards.ts`) with live status; clicking opens a **slide-over drawer** with a **structured outcome** (summary, findings, artifacts) and a collapsible read-only transcript (`src/ui/sub-agent-drawer.ts`, `src/styles/sub-agent-drawer.css`). While a run is active, `src/agents/sub-agent-runner.ts` pushes transcript snapshots through `onMessagesChange` (system/user seed, streaming assistant deltas, tool rounds); `src/agents/controller/controller.ts` (re-exported from `orchestrator.ts`) copies them onto `run.messages` and emits `sub-agent-events`; the open drawer re-renders on each emit. **`list_sub_agents`** / **`get_sub_agent_status`** are **session-scoped** (any parent turn in the same chat; merged with `chat.subAgentRuns` after reload). Terminal runs are copied into `chat.subAgentRuns` (`PersistedSubAgentRun[]` in `src/types.ts`, including `structuredOutcome` / `budgetEvents`) via `src/state/sub-agent-session-sync.ts` so the drawer works after reload. Spawn rows anchor after the parent tool bubble using `data-tool-call-id` on `.tool-call-msg` (`src/tools/loop.ts`).
 
 **Event-driven completion push (Build / General / Plan / Reef):** `spawn_sub_agent` defaults to **`wait: false`**. When a run settles, [`sub-agent-completion-push.ts`](../src/agents/sub-agent-completion-push.ts) coalesces pending completions and resumes the parent via `resumeParentChatWithMessage` (`suppressUserEcho: true`) once the chat is idle (`notifyChatStreamEnded` from [`loop.ts`](../src/tools/loop.ts)). Copy: [`sub-agent-resume-message.ts`](../src/chat/orchestrate/sub-agent-resume-message.ts). **Orchestrate** mode is excluded (supervisor owns stall recovery). Optional **check-in nudge** (`checkInNudgeMs`, default **120000**, Settings → Sub-agents; `0` disables) fires one non-cancelling parent note while a run is still active. Idempotent delivery tracks `deliveredRunIds`; transport errors retry once.
 
@@ -884,7 +884,8 @@ Parent tool loop can spawn **isolated sub-agents** (separate messages, model, to
 |---------|----------|
 | Types | `src/agents/types.ts` |
 | Config merge | `src/agents/sub-agent-config.ts`, `src/agents/defaults/sub-agents.json` |
-| Orchestrator | `src/agents/orchestrator.ts` — spawn, cancel, queue, `restartSubAgent`, `cancelAllForParentTurn`, `parentChatRuns` session index, list/status helpers; `deriveSubAgentTerminalReason` + `terminalReason` on `get_sub_agent_status` / aggregate JSON |
+| Controller | `src/agents/controller/{controller,scheduler,registry,watchdog}.ts` — spawn, cancel, queue, `restartSubAgent`, `cancelAllForParentTurn`, parent-chat/turn indexes, list/status helpers; `deriveSubAgentTerminalReason` + `terminalReason` on `get_sub_agent_status` / aggregate JSON. **`src/agents/orchestrator.ts`** is a thin re-export shim (naming maps to the prompt's "controller"). **MIN-140** Phase 3 adds heartbeat/watchdog supervision in `watchdog.ts`. |
+| Repetition detector | `src/agents/self-healing/detector.ts` — pure `detectRepetition` (reused by future watchdog) |
 | Completion push | `src/agents/sub-agent-completion-push.ts` — terminal delivery + check-in nudge (non-orchestrate) |
 | Events | `src/agents/sub-agent-events.ts` |
 | Runner | `src/agents/sub-agent-runner.ts` — headless tool loop; global `maxToolTurns` from sub-agents config; context budget enforcement; **`resolveSamplerPreset({ kind: 'sub-agent' })`** (not parent drawer temperature; default `max_tokens` 2048); optional `response_format` on final JSON turn when probed; content-JSON tool merge; tool nudge before premature finalization; structured JSON final turn; accumulates `usage` / `stats` on `SubAgentRun` for Orchestrate metrics rollup (MIN-36) |
